backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
web3==6.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.28.2
aiofiles==23.1.0
ipfshttpclient==0.7.0
starlette==0.26.1
pymongo==4.3.3 